Format "visualization" APIs
